Fix line drawing issue with flat brushes, fixes #16
Deleted two lines, I've tested it with all kinds of shapes and lines and it seems to function just fine without it. It's most likely an optimization intended to reduce the amount of particles drawn when drawing lines. The existing fill/no-fill logic should be good enough though, that logic makes it only draw an outline of the brush for each attempt in the line, instead of trying to refill the entire circle.