Optimize undo history memory usage (tpt++/LBPHacker eac92d1b04a5)
This commit includes a ton of other changes necessary to make the deltas work properly.
ElementDataContainer is now stored in a unique_ptr. There were issues with duplicate pointer references, and rather than solve them, using unique_ptr ensures pointer safety. (most likely explanation for the problem is that Snapshot was being copy constructed and copying the pointer to the element data, but I only figured this out after unique_ptr made me)
As part of this, every site that uses ElementDataContainer has also been updated. It's slightly more verbose now, but only because I used static_cast instead of a C-style cast

Something not in the tpt++ commit is the ability to run delta updates on element data containers. This is done for two elements, PRTI and LIFE, that store a ton of data. PRTI does a proper delta update on portals like tpt++. LIFE does it because gol is a very large member of the data container, but has absolutely no need to be restored in snapshots. It could have been done for ANIM too, but ANIM is a mod element and I don't want to put too much effort into it. Instead, ANIM will just not save element data at all if it's not present in the sim. Which will be true 99.99% of the time, and the few who do use ANIM won't ever notice the bugs (the old code also checked elementCount and nobody ever noticed)

Last minor thing is that signs was changed to a vector of signs, rather than a vector of sign pointers that need memory management. This was annoying and unnecessary.