Implement interface api from tpt++
A few new things were added to some components to handle all the required api functions. Also, added an option to Labels to not cut off text, required to match vanilla.
Most scripts look the same, but some are a little "off". Label workarounds fix most of this (I ignore the sizeX and sizeY parameters and calculate a new position), the remaining issues are mostly things calling legacy dialog functions, which still use the legacy dialogs in my mod (because they need to be blocking), and also scripts like cracker1000's that draw graphics outside of the window bounds